# docs/swagger.yaml
basePath: /im
definitions:
...
  enums.UserStatus:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-comments:
      UserStatusActive: 正常狀態
      UserStatusBlocked: 被封鎖
      UserStatusInactive: 無效或未激活
    x-enum-varnames:
    - UserStatusActive
    - UserStatusBlocked
    - UserStatusInactive
...
  req.GroupMembersCreate:
    type: object
  req.GroupMembersDelete:
    properties:
      id:
        type: string
    type: object
  req.GroupMembersUpdate:
    type: object
  req.GroupsCreate:
    properties:
      description:
        description: 描述
        type: string
      group_name:
        description: 群組名稱
        type: string
    required:
    - group_name
    type: object
  req.GroupsDelete:
    properties:
      id:
        type: string
    type: object
  req.GroupsGet:
    type: object
  req.GroupsGetList:
    properties:
      index:
        description: 頁碼
        type: integer
      order:
        description: 排序
        example: id asc
        type: string
      size:
        description: 筆數
        type: integer
    type: object
  req.GroupsUpdate:
    type: object
...
  req.UsersCreate:
    properties:
      email:
        type: string
      nickname:
        type: string
      password:
        minLength: 6
        type: string
      phone_number:
        type: string
      username:
        minLength: 6
        type: string
    required:
    - email
    - password
    - phone_number
    - username
    type: object
  req.UsersGet:
    properties:
      email:
        type: string
      nickname:
        type: string
      password:
        type: string
      phone_number:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/enums.UserStatus'
        description: 用戶狀態
      username:
        type: string
    type: object
  req.UsersGetList:
    properties:
      index:
        description: 頁碼
        type: integer
      order:
        description: 排序
        example: id asc
        type: string
      size:
        description: 筆數
        type: integer
    type: object
  req.UsersLogin:
    properties:
      password:
        description: 密碼
        minLength: 6
        type: string
      username:
        description: 使用者名稱
        minLength: 6
        type: string
    required:
    - password
    - username
    type: object
  req.UsersUpdate:
    properties:
      email:
        type: string
      id:
        type: string
      nickname:
        type: string
      phone_number:
        type: string
      username:
        minLength: 6
        type: string
    required:
    - email
    - id
    - phone_number
    - username
    type: object
...
  resp.APIResponse-resp_GroupMembersGetList:
    properties:
      code:
        description: 回傳代碼
        type: string
      data:
        allOf:
        - $ref: '#/definitions/resp.GroupMembersGetList'
        description: 資料
      msg:
        description: 訊息
        type: string
    type: object
  resp.APIResponse-resp_GroupsGet:
    properties:
      code:
        description: 回傳代碼
        type: string
      data:
        allOf:
        - $ref: '#/definitions/resp.GroupsGet'
        description: 資料
      msg:
        description: 訊息
        type: string
    type: object
  resp.APIResponse-resp_GroupsGetList:
    properties:
      code:
        description: 回傳代碼
        type: string
      data:
        allOf:
        - $ref: '#/definitions/resp.GroupsGetList'
        description: 資料
      msg:
        description: 訊息
        type: string
    type: object
...
  resp.APIResponse-resp_UsersGet:
    properties:
      code:
        description: 回傳代碼
        type: string
      data:
        allOf:
        - $ref: '#/definitions/resp.UsersGet'
        description: 資料
      msg:
        description: 訊息
        type: string
    type: object
  resp.APIResponse-resp_UsersGetList:
    properties:
      code:
        description: 回傳代碼
        type: string
      data:
        allOf:
        - $ref: '#/definitions/resp.UsersGetList'
        description: 資料
      msg:
        description: 訊息
        type: string
    type: object
  resp.APIResponse-resp_UsersLogin:
    properties:
      code:
        description: 回傳代碼
        type: string
      data:
        allOf:
        - $ref: '#/definitions/resp.UsersLogin'
        description: 資料
      msg:
        description: 訊息
        type: string
    type: object
  resp.APIResponse-string:
    properties:
      code:
        description: 回傳代碼
        type: string
      data:
        description: 資料
        type: string
      msg:
        description: 訊息
        type: string
    type: object
...
  resp.GroupMembersGet:
    properties:
      custom_permissions:
        type: string
      group_id:
        type: string
      invited_by_code:
        type: string
      invited_by_user_id:
        type: string
      joined_at:
        type: string
      last_seen:
        type: string
      message_read_up_to:
        type: integer
      mute_notifications:
        type: boolean
      nickname_in_group:
        type: string
      role:
        type: string
      status_in_group:
        type: string
      user_id:
        type: string
      user_name:
        type: string
    type: object
  resp.GroupMembersGetList:
    properties:
      data:
        items:
          $ref: '#/definitions/resp.GroupMembersGet'
        type: array
      page:
        $ref: '#/definitions/resp.PageResponse'
    type: object
  resp.GroupsGet:
    properties:
      description:
        type: string
      group_name:
        type: string
      id:
        type: string
    type: object
  resp.GroupsGetList:
    properties:
      data:
        items:
          $ref: '#/definitions/resp.GroupsGet'
        type: array
      page:
        $ref: '#/definitions/resp.PageResponse'
    type: object
...
  resp.PageResponse:
    properties:
      index:
        description: 頁碼
        type: integer
      pages:
        description: 總頁數
        type: integer
      size:
        description: 筆數
        type: integer
      total:
        description: 總筆數
        type: integer
    type: object
  resp.UsersGet:
    properties:
      email:
        description: 電子郵件地址
        type: string
      id:
        description: uid
        type: string
      nickname:
        description: 用戶暱稱
        type: string
      password:
        description: 密碼
        type: string
      phone_number:
        description: 手機號碼
        type: string
      username:
        description: 用戶名稱
        type: string
    type: object
  resp.UsersGetList:
    properties:
      data:
        items:
          $ref: '#/definitions/resp.UsersGet'
        type: array
      page:
        $ref: '#/definitions/resp.PageResponse'
    type: object
  resp.UsersLogin:
    properties:
      token:
        description: 登入Token
        type: string
      username:
        description: 用戶名稱
        type: string
    type: object
...
host: localhost:9000
info:
  contact:
    name: API Support
  description: This is a project im.
  termsOfService: http://swagger.io/terms/
  title: Im
  version: "1.0"
paths:
...
  /group:
    delete:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.GroupsDelete'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-string'
      summary: Delete
      tags:
      - groups
    get:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.GroupsGetList'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-resp_GroupsGetList'
      summary: GetList
      tags:
      - groups
    post:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.GroupsCreate'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-string'
      summary: 創建群組
      tags:
      - groups
    put:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.GroupsUpdate'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-string'
      summary: Update
      tags:
      - groups
...
  /group-members:
    delete:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.GroupMembersDelete'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-string'
      summary: Delete
      tags:
      - groupMembers
    post:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.GroupMembersCreate'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-string'
      summary: Create
      tags:
      - groupMembers
    put:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.GroupMembersUpdate'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-string'
      summary: Update
      tags:
      - groupMembers
  /group-members/:id:
    get:
      parameters:
      - description: 群組ID
        in: query
        name: id
        type: string
      - description: 角色
        enum:
        - owner
        - admin
        - member
        in: query
        name: role
        type: string
        x-enum-varnames:
        - GroupRoleOwner
        - GroupRoleAdmin
        - GroupRoleMember
      - description: 群組內狀態
        in: query
        name: status_in_group
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-resp_GroupMembersGetList'
      summary: GetList
      tags:
      - groupMembers
  /group/:id:
    get:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.GroupsGet'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-resp_GroupsGet'
      summary: Get
      tags:
      - groups
...
  /users:
    put:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.UsersUpdate'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-string'
      summary: 用戶訊息修改
      tags:
      - users
  /users/:id:
    get:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.UsersGet'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-resp_UsersGet'
      summary: 取得用戶訊息
      tags:
      - users
...
  /users/login:
    post:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.UsersLogin'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-resp_UsersLogin'
      summary: 用戶登錄並返回授權令牌
      tags:
      - users
  /users/logout:
    post:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-string'
      security:
      - ApiKeyAuth: []
      summary: 用戶登出
      tags:
      - users
  /users/register:
    post:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.UsersCreate'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-string'
      summary: 用戶註冊
      tags:
      - users
  /users/search:
    get:
      parameters:
      - description: param
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/req.UsersGetList'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.APIResponse-resp_UsersGetList'
      summary: 用戶清單
      tags:
      - users
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: token
    type: apiKey
swagger: "2.0"